Change -r to -i and support loading files before REPL.
This change slightly modifies how the REPL starts. Now
the run has a repl: boolean setting, instead of using
a target. This means the REPL can just process the file
targets like anything else when it launches.

If we can reserve a magic character in the REPL (e.g. %)
we could support something like:

   %load test.em

...during REPL sessions, which would import the file and
the drop control back to the user. I feel like this would
be really nice.

Overall, the code is still quite ugly, but it works, so
I figured I'd commit here. Feedback welcome.