Some gestures towards usability.
The REPL starts with a banner, produces a prompt (>>>),
and has an incredibly rudimentary system of meta commands.

Right now, %help and %quit are grabbed by the REPL, to
produce help output or quit, respectively. Eventually, my
goal is to support other more interesting commands, to
inspect the scope, look up names, do search/completion,
and so on.

There's an argument to be made that maybe % commands
should be implemented in Emily. I think this is a good
idea but right now it's not clear most of the commands
will have good implementations. I guess it depends on
how what kind of introspection Emily offers (something
I don't know a lot about yet).